sales-service/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
motor==2.5.1
pydantic==1.8.2
python-dotenv==0.19.0
httpx==0.23.0
pytest==6.2.5
pytest-asyncio==0.16.0
pytest-cov==2.12.1 